requirements.txt
# Flask
requests

Flask==2.1.1
google-auth==2.6.0
google-auth-oauthlib==0.4.6
google-auth-httplib2==0.1.0
google-api-python-client==2.35.0
python-dotenv==0.20.0

